Showcase reproducibility issue across machines
The problem stems from the permission bits set on host being preserved both when files are copied into the docker image and when they are added to the tar archive.

`reproducible-package` reproduces because the umask on my machine does not result in different permissions.

`shell-script-example` on the other hand does not reproduce because the shell script gets permissions `lrwxr-xr-x` on my MacOS host which are different than `lrwxrwxrwx` assigned on the github runner.